refactor(backend): extract sales analytics report service
