+++ backend/src/modules/sales/sales.module.ts
@@ -31,6 +31,7 @@ import { SalesOrderService } from './services/sales-order.service'; // Temporari
 import { InvoiceService } from './services/invoice.service';
 import { PaymentService } from './services/payment.service';
 import { SalesAnalyticsService } from './services/sales-analytics.service';
+import { SalesAnalyticsReportService } from './services/sales-analytics-report.service';
 import { InventoryIntegrationService } from './services/inventory-integration.service';
 import { SalesOrderFulfillmentService } from './services/sales-order-fulfillment.service';
 import { SalesOrderLifecycleService } from './services/sales-order-lifecycle.service';
@@ -71,6 +72,7 @@ import { TransactionManager } from '../../common/utils/transaction.util';
     InvoiceService,
     PaymentService,
     SalesAnalyticsService,
+    SalesAnalyticsReportService,
     InventoryIntegrationService,
     SalesOrderFulfillmentService,
     SalesOrderLifecycleService,
